deprecated s:GitStat() in honer of improved s:GitDiff()/s:GitShow()
Added new param to s:GitDiff() (a:4) and s:GitShow() (a:1) to pass a set
of diff modes used to display diff information. This param is a string,
each char in it is a diff view to include into git-diff output:

    p => -p
    s => --stat
    d => --dirstat
    r => --raw
    n => --numstat
    o => --shortstat

Besides this --stat param is passed as --stat=0,1000 to avoid
abbreviated file names, so "gf" should work anyway.